add typing for nodes
